Refactor topics packages
